ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pdfminer.six==20221105
spacy==3.7.4
python-multipart==0.0.9
pydantic==2.7.1
